+++ RELEASE_NOTES.md
@@ -7,6 +7,8 @@ _____
 * Update GTDB data to r214.1
 * Removed full_tree option (requires nodes with too much memory)
 * Added db_ver argument to allow either r207 or r214 refdata
+* Update Mash to v2.3
+* Generate mash_db for species rep genomes and add to refdata
 
 1.3.0
 _____